Removing WeblogEntry interface, fixing 1.4.2 compile problems
git-svn-id: https://svn.apache.org/repos/asf/incubator/roller/branches/roller_3.0@412521 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/roller/business/hibernate/HibernateUserManagerImpl.java
@@ -921,7 +921,7 @@ public class HibernateUserManagerImpl implements UserManager {
                     (String)row[2], 
                     (String)row[3], 
                     "statCount.weblogCommentCountType", 
-                    new Long((Integer)row[0]).longValue()));
+                    new Long(((Integer)row[0]).intValue()).longValue()));
             }
             return results;
         } catch (Throwable pe) {
